# MemoryKnowledge/openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge Service API
  description: |
    Standalone knowledge service for Code-Graph + LLM-Wiki.
    Provides 28 endpoints for managing and querying code repositories
    and wiki knowledge bases.

    ## Multi-tenancy (001)

    Every endpoint requires the tenant identity `service_id` in the
    **`x-tdai-service-id` request header** (`^[A-Za-z0-9_-]+$`, missing/malformed → 400).
    This is the same value the kernel uses for store/VDB routing, so tenant identity is
    unified end-to-end. All lookups are scoped by `service_id`; an id-only endpoint
    targeting another tenant's resource returns 404. `service_id` is never a request-body
    field and is not echoed in responses.

    ## Asset ID Model

    Assets have globally unique, immutable IDs:

    | Asset Type  | Field           | Prefix   | Example        |
    |-------------|-----------------|----------|----------------|
    | LLM-Wiki    | `wiki_id`       | `wiki-`  | `wiki-9c1f2b`  |
    | Code-Graph  | `code_graph_id` | `cg-`    | `cg-7e3d10`    |

    ## Response Envelope

    All responses use `ApiResponseEnvelope`:
    - `code: 0` means success; non-zero means business error
    - `data` contains the payload on success, `null` on error
    - `message` provides human-readable status

  version: "1.0.0"
  contact:
    name: Knowledge Service
servers:
  - url: http://localhost:8421/v3
    description: Local development server
...
paths:
...
  /wiki/create:
    post:
      tags: [LLM-Wiki]
      summary: Create wiki metadata + directory shell
      description: |
        Generates wiki_id (wiki-{8char}), creates DB record + directory shell.
        Idempotent: same (service_id, team_id, name) → returns existing (200).
        Does not auto-trigger ingest; call /wiki/ingest explicitly.
      operationId: createWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WikiDetail'
        '200':
          description: Idempotent hit
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WikiDetail'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/get:
    post:
      tags: [LLM-Wiki]
      summary: Get wiki metadata by ID
      operationId: getWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiGetRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WikiDetail'
        '404':
          description: Wiki not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/list:
    post:
      tags: [LLM-Wiki]
      summary: List wikis for a team
      operationId: listWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiListRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WikiListData'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/ingest:
    post:
      tags: [LLM-Wiki]
      summary: Trigger LLM ingest (async)
      description: |
        Background SerialQueue processes raw/sources/* via LLM to generate
        wiki pages + BM25 index. Returns immediately with status=pending.
      operationId: ingestWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiIngestRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          wiki_id:
                            type: string
                          status:
                            type: string
        '404':
          description: Wiki not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/delete:
    post:
      tags: [LLM-Wiki]
      summary: Batch delete wikis
      description: |
        Deletes wikis in any status, including `pending`/`processing` (008): a
        delete never returns 409. In-flight ingest is signalled to abort at its
        next checkpoint; the row is hard-deleted and connection/disk/index
        resources are cleaned up. Cross-memory ids report as "not found".
      operationId: deleteWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiDeleteRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BatchDeleteResult'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/raw/ls:
    post:
      tags: [LLM-Wiki]
      summary: List raw source files
      operationId: rawLsWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiRawLsRequest'
      responses:
        '200':
          description: Success
        '404':
          description: Wiki not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/raw/read:
    post:
      tags: [LLM-Wiki]
      summary: Batch read raw source files
      operationId: rawReadWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiRawReadRequest'
      responses:
        '200':
          description: Success
        '400':
          description: Invalid path or too many files
        '404':
          description: Wiki not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/raw/write:
    post:
      tags: [LLM-Wiki]
      summary: Batch write raw source files
      operationId: rawWriteWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiRawWriteRequest'
      responses:
        '200':
          description: Success
        '400':
          description: Invalid path, too large, or processing
        '404':
          description: Wiki not found
        '409':
          description: Wiki is processing
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/raw/rm:
    post:
      tags: [LLM-Wiki]
      summary: Batch delete raw source files + cascade
      operationId: rawRmWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiRawRmRequest'
      responses:
        '200':
          description: Success
        '400':
          description: Invalid path or too many files
        '404':
          description: Wiki not found
        '409':
          description: Wiki is processing
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/page/ls:
    post:
      tags: [LLM-Wiki]
      summary: List wiki pages (processed)
      operationId: pageLsWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiPageLsRequest'
      responses:
        '200':
          description: Success
        '404':
          description: Wiki not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/page/read:
    post:
      tags: [LLM-Wiki]
      summary: Batch read wiki pages
      operationId: pageReadWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiPageReadRequest'
      responses:
        '200':
          description: Success
        '400':
          description: Invalid path or too many refs
        '404':
          description: Wiki not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/page/write:
    post:
      tags: [LLM-Wiki]
      summary: Batch write wiki pages (auto-injects locked:true)
      operationId: pageWriteWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiPageWriteRequest'
      responses:
        '200':
          description: Success
        '400':
          description: Invalid path, forbidden path, or too large
        '404':
          description: Wiki not found
        '409':
          description: Wiki is processing
        '413':
          description: Content exceeds size limit
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/page/rm:
    post:
      tags: [LLM-Wiki]
      summary: Batch delete wiki pages + cascade
      operationId: pageRmWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiPageRmRequest'
      responses:
        '200':
          description: Success
        '400':
          description: Invalid path or forbidden path
        '404':
          description: Wiki not found
        '409':
          description: Wiki is processing
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/graph:
    post:
      tags: [LLM-Wiki]
      summary: Get wiki knowledge graph (nodes, edges, communities)
      operationId: graphWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiGraphRequest'
      responses:
        '200':
          description: Success
        '404':
          description: Wiki not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /wiki/search:
    post:
      tags: [LLM-Wiki]
      summary: Search wiki pages (BM25 full-text)
      operationId: searchWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiSearchRequest'
      responses:
        '200':
          description: Success
        '400':
          description: Missing query
        '404':
          description: Wiki not found
        default:
          $ref: '#/components/responses/ErrorResponse'

  # ═══════════════════════════════════════════════
  # Code-Graph (13 endpoints)
  # ═══════════════════════════════════════════════
...
components:
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseEnvelope'
  schemas:
    ApiResponseEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          description: 0 = success, non-zero = business error
        message:
          type: string
        request_id:
          type: string
        data:
          nullable: true
...
    WikiDetail:
      type: object
      properties:
        wiki_id:
          type: string
        team_id:
          type: string
        name:
          type: string
        service_url:
          type: string
          nullable: true
          description: Tools self-discovery base URL (includes API prefix, e.g. http://host:8421/v3)
        summary:
          type: string
          nullable: true
          maxLength: 256
        status:
          type: string
          enum: [draft, pending, processing, ready, failed]
          description: draft = 建壳未加工；pending/processing = 摄入中；ready = 最近一次摄入成功；failed = 可重试。已有已发布内容在排队、摄入和失败期间仍可查询，新内容在摄入提交后可见。
        sync_error:
          type: string
          nullable: true
        version:
          type: string
        owner_user_id:
          type: string
          nullable: true
        page_count:
          type: integer
          nullable: true
        last_sync_at:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
    WikiCreateRequest:
      allOf:
        - $ref: '#/components/schemas/IdFields'
        - type: object
          required: [team_id, name]
          properties:
            name:
              type: string
...